Added logic to checkout shopping cart

+++ src/main/java/com/sagar/solutions/storefront/domain/shoppingcart/ShoppingCart.java
@@ -11,7 +11,7 @@ public class ShoppingCart {
 
     private CartStatus cartStatus;
 
-    private Integer purchaseOrderId;
+    private Long purchaseOrderId;
